+++ src/grokcore/chameleon/components.py
@@ -15,6 +15,7 @@
 
 import os
 import sys
+import typing
 
 import chameleon.i18n
 import martian
@@ -67,7 +68,7 @@ class PageTemplate(PageTemplate):
                     target_language=None, default=None):
                 # We swap context with the request, that is required for
                 # zope.i18n.translate.
-                if isinstance(msgid, str):
+                if isinstance(msgid, typing.Hashable):
                     return zope.i18n.translate(
                         msgid, domain, mapping, request, target_language,
                         default)