+++ CHANGES.rst
@@ -8,7 +8,7 @@ CHANGES
   allowing to use this version.
 
 - Update translation handling to be compatible with Chameleon 4.3+: do not
-  break no unhashable message ids, just return them.
+  break on unhashable message ids, just return them.
 
 
 4.0 (2023-02-09)